// store/modules/menu/types.ts
export enum MenuType {
  /**
      @types Action
    **/
  ACTION_GET_MENUS = 'ACTION_GET_MENUS',

  /**
      @types Mutation
    **/
  MUTATION_SET_MENUS = 'MUTATION_SET_MENUS',
  MUTATION_SET_CHANGE_STATUS_TOGGLE = 'MUTATION_SET_CHANGE_STATUS_TOGGLE',

  /**
      @types State
    **/
  STATE_LIST_MENUS = 'STATE_LIST_MENUS',
}